requirements.txt
Pillow==10.1.0
vk-api==11.9.9